Move request complexity into its own file
