+++ lib/github_integration/install-branch.js
@@ -1,4 +1,4 @@
-const da = require('../utils/docassemble_api.js');
+const dai = require('./docassemble_api_interface');
 const log = require('../utils/log');
 const session_vars = require('../utils/session_vars' );
 const time = require('../utils/time');
@@ -12,7 +12,7 @@ const setup = async () => {
   log.info({ type: `setup`, pre: `Trying to create a docassemble project.` });
   await create_project();
   log.info({ type: `setup`, pre: `Trying to pull code from the "${ session_vars.get_branch_name() }" branch.` });
-  let task_id = await da.pull();
+  let task_id = await dai.pull();
   await wait_for_server_to_restart( task_id );
   
   return true;
@@ -23,7 +23,7 @@ const takedown = async () => {
   *   DOCASSEMBLE_DEVELOPER_API_KEY secret, delete the
   *   Project created for this test.
   *   For local development, clean up project name file. */
-  await da.delete_project();
+  await dai.delete_project();
   session_vars.delete_project_name();
 };
 
@@ -46,7 +46,7 @@ const create_project = async function () {
 
       project_name = session_vars.get_project_base_name() + name_incrementor;
       log.info({ type: `setup`, pre: `Trying to create the project with the name ${ project_name }` });
-      response = await da.create_project( project_name );
+      response = await dai.create_project( project_name );
       result = response.data;
 
     } catch ( error ) {
@@ -97,7 +97,7 @@ const wait_for_server_to_restart = async ( task_id ) => {
     num_tries++;
 
     try {
-      done_restarting = await da.has_server_restarted( task_id );
+      done_restarting = await dai.has_server_restarted( task_id );
     } catch ( error ) {
       log.debug({ pre: `Errored while trying to see if the server restarted.` });
       console.log( error );
